수정

+++ EngineNiaServer2021/NiaIpSdnSflowLinkageApplication/src/main/java/com/nia/ip/sdn/sflow/linkage/mapper/SflowMapper.java
@@ -8,6 +8,6 @@ import java.util.HashMap;
 
 @Mapper
 public interface SflowMapper {
-    int selectSflowSeq();
+    long selectSflowSeq();
     void insertSflowData(SflowCollectVo sflowCollectVo);
 }
